Replace timezone.js with timezone-list.js
