+++ api/commands/misc/timezone.js
@@ -3,18 +3,37 @@ const exceptions  = require('../../util/exceptions.json');
 const insertTz    = require('../../../db/queries/insertTimezone.js');
 const moment      = require('moment-timezone');
 
+const timezones = {
+  1: 'America/Puerto_Rico',
+  2: 'America/New_York',
+  3: 'America/Chicago',
+  4: 'America/Denver',
+  5: 'America/Phoenix',
+  6: 'America/Los_Angeles',
+  7: 'America/Anchorage',
+  8: 'Pacific/Honolulu'
+};
+
 const command = {
   name: 'timezone',
   group: 'misc',
   memberName: 'timezone',
-  description: 'Set your timezone. To view a list of zones, please visit ' +
-    '`https://en.wikipedia.org/wiki/List_of_tz_database_time_zones`',
-  examples: ['rbot timezone America/Los_Angeles'],
+  description: 'Select from a list of timezones.',
   args: [
     {
       key: 'content',
-      prompt: 'What is your timezone?',
-      type: 'string'
+      prompt: 'Quick-select your timezone by entering a number ' +
+        'from the following list. Otherwise, enter the name of your zone ' +
+        'if it is not included in the following list. ' +
+        '[1] Atlantic ' +
+        '[2] Eastern ' +
+        '[3] Central ' +
+        '[4] Mountain ' +
+        '[5] MST ' +
+        '[6] Pacific ' +
+        '[7] Alaska ' +
+        '[8] Hawaii',
+      type: 'number'
     }
   ]
 };
@@ -25,9 +44,12 @@ module.exports = class TimezoneCommand extends Command {
   }
 
   run(msg, { content }) {
-    if (!moment.tz.names().includes(content)) {
+    if (
+      !Object.prototype.hasOwnProperty.call(timezones, content) &&
+      !moment.tz.names().includes(content)
+    ) {
       return new Promise((resolve, reject) =>
-        reject(msg.say(exceptions.invalid_timezone)));
+        reject(msg.say(exceptions.invalid_timezone_new)));
     }
 
     return insertTz([
